server/requirements.txt
flask
flask-cors
langchain
google-generativeai
langchain-google-genai
# --upgrade langchain-google-genai
python-dotenv